Refactor hodlr utils from expect utils
